requirements.txt
django-adminlte3